migrate in a new chat application
