+++ project/carpool/views.py
@@ -7,9 +7,38 @@ from carpool.forms import CreateRideForm
 from carpool.tasks import get_autocompletion, get_routing
 from asgiref.sync import sync_to_async
 from django.db.models.functions import TruncDate
-
+from django.http import HttpResponse
+from django.views.decorators.http import require_http_methods
+from chat.models import ChatRequest
 
 from django.core.paginator import Paginator
+
+
+@login_required
+@require_http_methods(["POST"])
+def change_jrequest_status(request, jr_pk):
+    join_request = get_object_or_404(ChatRequest, pk=jr_pk)
+
+    if request.user != join_request.ride.driver:
+        return HttpResponse("You are not the driver of this post", status=403)
+
+    action = request.POST.get("action")
+    if action == "accept":
+        join_request.status = ChatRequest.Status.ACCEPTED
+        # Add the user to the ride
+        join_request.ride.rider.add(join_request.user)
+
+    elif action == "decline":
+        join_request.status = ChatRequest.Status.DECLINED
+        # Remove the user to the ride if they were added
+        if join_request.user in join_request.ride.rider.all():
+            join_request.ride.rider.remove(join_request.user)
+
+    else:
+        return HttpResponse("Invalid action", status=400)
+
+    join_request.save()
+    return redirect("chat:index")
 
 
 @login_required
