+++ project/chat/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from chat.models import ChatMessage, ChatRequest
+
+admin.site.register(ChatMessage)
+admin.site.register(ChatRequest)
