+++ project/project/asgi.py
@@ -13,7 +13,7 @@ from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
 from channels.security.websocket import AllowedHostsOriginValidator
 from django.core.asgi import get_asgi_application
-from carpool import routing as carpool_routing
+from chat import routing
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "project.settings")
 
@@ -22,7 +22,7 @@ application = ProtocolTypeRouter(
         "http": get_asgi_application(),
         # Add other protocols here if needed, e.g., WebSocket
         "websocket": AllowedHostsOriginValidator(
-            AuthMiddlewareStack(URLRouter(carpool_routing.websocket_urlpatterns)),
+            AuthMiddlewareStack(URLRouter(routing.websocket_urlpatterns)),
         ),
     }
 )
